api_yamdb/requirements.txt
Django==3.0.5
django-filter==21.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.1.0
PyJWT==2.3.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python-dotenv==0.19.2
requests==2.26.0
gunicorn==20.0.4
psycopg2-binary==2.8.6
